Door rework, Key label updated
I have reworked the door to the specifications of the diagrams to make
sure to use the Goal class. I have also updated the key label to
properly display when LevelCompleted is called. THIS UPDATE NEEDS
gameBoard.levelCompleted(); PLACED WITHIN THE COLLISION DETECTOR BASED
ON PLAYER LOCATION IN RELATION TO THE DOOR LOCATION.